feat(generic_bounds): use `AsyncFnOnce` for async upgrade needed callback
BREAKING CHANGE: `AsyncFnOnce` was stabilized in version 1.85.0 of Rust,
so the MSRV had to be bumped to accommodate this. Additionally,
synchronous closures that use `async move` blocks are not drop-in
replacements for `async` closures, so users of
`OpenDbRequestBuilder::with_on_upgrade_needed_fut` will have to make minor
changes to their invocations.