feat(database): provide active transaction (instead of database) to async upgrade needed callback
BREAKING CHANGE: The callback provided to
`OpenDbRequestBuilder::with_on_upgrade_needed_fut` must now accept a
reference to a `Transaction` rather than a `Database`. For most cases,
existing code can be made compatible by calling `Transaction::database`
in order to get access to the underlying database. The caveat is that
the `Transaction::database` provides a reference, rather than ownership
of the `Database`.